requirements.txt
fastapi
uvicorn
mcp
httpx
pytest
pytest-asyncio
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
